% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloch_axisChrono.R
\name{axisChrono}
\alias{axisChrono}
\title{Axis for time-calibrated phylogenies}
\usage{
axisChrono(side = 1, unit = NULL, fact = 1, ...)
}
\arguments{
\item{side}{A numeric value specifying the side where the axis is plotted:
1: below, 2: left, 3: above, 4: right.}

\item{unit}{A character string used to provide information on time units; defaults to "Ma" (million years ago).}

\item{fact}{A real number used to scale the axis.}

\item{...}{
  Arguments passed on to \code{\link[graphics:axis]{graphics::axis}}
  \describe{
    \item{\code{at}}{the points at which tick-marks are to be drawn.  Non-finite
    (infinite, \code{NaN} or \code{NA}) values are omitted.  By default
    (when \code{NULL}) tickmark locations are computed, see
    \sQuote{Details} below.}
    \item{\code{labels}}{this can either be a logical value specifying whether
    (numerical) annotations are to be made at the tickmarks, or a
    character or expression vector of labels to be placed at the
    tickpoints.  (Other objects are coerced by \code{\link{as.graphicsAnnot}}.)
    If this is not logical, \code{at} should also be supplied and of the
    same length.  If \code{labels} is of length zero after coercion,
    it has the same effect as supplying \code{TRUE}.}
    \item{\code{tick}}{a logical value specifying whether tickmarks and an axis line
    should be drawn.}
    \item{\code{line}}{the number of lines into the margin at which the axis line
    will be drawn, if not \code{NA}.}
    \item{\code{pos}}{the coordinate at which the axis line is to be drawn:
    if not \code{NA} this overrides the value of \code{line}.}
    \item{\code{outer}}{a logical value indicating whether the axis should be
    drawn in the outer plot margin, rather than the standard plot
    margin.}
    \item{\code{font}}{font for text.  Defaults to \code{par("font")}.}
    \item{\code{lty}}{line type for both the axis line and the tick marks.}
    \item{\code{lwd}}{line widths for the axis line and the tick
    marks.  Zero or negative values will suppress the line or ticks.}
    \item{\code{lwd.ticks}}{line widths for the axis line and the tick
    marks.  Zero or negative values will suppress the line or ticks.}
    \item{\code{col}}{colors for the axis line and the tick marks
    respectively.  \code{col = NULL} means to use \code{par("fg")},
    possibly specified inline, and \code{col.ticks = NULL} means to use
    whatever color \code{col} resolved to.}
    \item{\code{col.ticks}}{colors for the axis line and the tick marks
    respectively.  \code{col = NULL} means to use \code{par("fg")},
    possibly specified inline, and \code{col.ticks = NULL} means to use
    whatever color \code{col} resolved to.}
    \item{\code{hadj}}{adjustment (see \code{\link[graphics]{par}("adj")}) for all labels
    \emph{parallel} (\sQuote{horizontal}) to the reading direction.  If
    this is not a finite value, the default is used (centring for
    strings parallel to the axis, justification of the end nearest the
    axis otherwise).}
    \item{\code{padj}}{adjustment for each tick label \emph{perpendicular} to the
    reading direction.  For labels parallel to the axes, \code{padj = 0}
    means right or top alignment, and \code{padj = 1} means left or bottom
    alignment.  This can be a vector given a value for each string, and
    will be recycled as necessary.

    If \code{padj} is not a finite value (the default), the value of
    \code{par("las")} determines the adjustment.  For strings plotted
    perpendicular to the axis the default is to centre the string.}
    \item{\code{gap.axis}}{an optional (typically non-negative) numeric factor to
    be multiplied with the size of an \sQuote{m} to determine the
    minimal gap between labels that are drawn, see \sQuote{Details}.
    The default, \code{NA}, corresponds to \code{1} for tick labels drawn
    \emph{parallel} to the axis and \code{0.25} otherwise, i.e., the
    default is equivalent to \preformatted{  perpendicular <- function(side, las) {
    is.x <- (side \%\% 2 == 1) # is horizontal x-axis
    ( is.x && (las \%in\% 2:3)) ||
    (!is.x && (las \%in\% 1:2))
  }
  gap.axis <- if(perpendicular(side, las)) 0.25 else 1}% end{pre..}

    \code{gap.axis} may typically be relevant when \code{at = ..}
    tick-mark positions are specified explicitly.
  }
  }}
}
\description{
This function adds a time scale on the side of a phylogeny plot.
}
\author{
Christoph Heibl
}
