% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_ages.R
\name{plot_node_ages}
\alias{plot_node_ages}
\title{Plot a Chronogram and Add Node Age/Calibration Data to Nodes}
\usage{
plot_node_ages(
  phy,
  cex_tips = 1,
  calibration_summary,
  color_points,
  pch = 20,
  cex_pch = 1,
  color_bars = "#80808050",
  lty_bars = "solid",
  lwd_bars = 7
)
}
\arguments{
\item{phy}{A `phylo` object with branch length proportional to time}

\item{cex_tips}{A numeric value indicating **c**haracter **ex**pansion (i.e.,
size scaling factor) for `phy` tip labels. Default to 1.}

\item{calibration_summary}{An output of [datelife:::summary.matchedCalibrations()]}

\item{color_points}{A named vector of colors. Names must correspond to study names
in `calibration_summary$in_phy$references`.
If vector is not named, colors will be recycled.}

\item{pch}{A numeric vector indicating the type of point to plot.
See [graphics::par()] for options. Defaults to 20 = "bullet circle".}

\item{cex_pch}{A numeric value indicating **c**haracter **ex**pansion (i.e.,
size scaling factor) of node age point symbols defined by `pch`. Default to 1.}

\item{color_bars}{A character vector of one element indicating the color for
 node age distribution bars. Defaults to "#80808050", which is hex for gray
("#808080") with 50\% transparency.}

\item{lty_bars}{A numeric vector or character string indicating the plotting
line type for bars. See [graphics::par()] for options. Default to "solid" (1 if numeric).}

\item{lwd_bars}{A numeric vector indicating the plotting width for bars.
See [graphics::par()] for options. Default to 7.}
}
\description{
`plot_node_ages` plots a chronograms given in `phy` and adds age
  data [points()] from `calibration_summary` to the corresponding nodes.
}
