% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phylo_all}
\alias{plot_phylo_all}
\title{plot all trees with study titles and geochronological axis}
\usage{
plot_phylo_all(
  trees,
  cex = graphics::par("cex"),
  include = TRUE,
  individually = TRUE,
  write = "no",
  file = "phylo_all",
  plot_type = "phyloch",
  max_depth
)
}
\arguments{
\item{trees}{A list of trees as multiPhylo or as a plain list object.}

\item{cex}{a real number giving the \bold{c}haracter \bold{ex}pansion.}

\item{include}{Boolean or numeric vector. Default to TRUE, keep all chronograms
in trees. If FALSE, exclude chronograms with only two tips. If numeric, it is used
as indices to subset trees object.}

\item{individually}{Boolean indicating if trees should be plotted one by one or all on the same file}

\item{write}{A character vector of length 1. Use pdf or png to write a file
on those formats respectively. Anything else will not write any file.}

\item{file}{A character string giving the name and path to write the files to.}

\item{plot_type}{A character vector of length one indicating the type of
plot to generate. Options are:
\describe{
    \item{"phyloch"}{It uses the functions [ape::plot.phylo()] and [phyloch::axisGeo()].}
    \item{"strap"}{It uses the function [strap::geoscalePhylo()] from the package [strap].}
    \item{"phytools"}{Not implemented yet. It will use functions from the package [phytools]}
    }}

\item{max_depth}{A numeric vector of length 1, indicating the upper limit for
the time scale on the x axis for all plots. If none is provided, it is estimated
by getting the tree depth of the oldest chronogram, adds 5 and rounds to the
closest 10. See details.}
}
\description{
plot all trees with study titles and geochronological axis
}
\details{
Currently, max_depth is obtained with round(max(sapply(trees, function(x) max(ape::branching.times(x)))) + 5, digits = -1)
}
