% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{subset_trees}
\alias{subset_trees}
\title{subset chronograms for plotting densitree plots and phylo_all plots}
\usage{
subset_trees(chronograms, include = TRUE)
}
\arguments{
\item{chronograms}{A list of chronograms as multiPhylo or as a plain list object.}

\item{include}{Boolean or numeric vector. Default to TRUE, keep all chronograms
in chronograms. If FALSE, exclude chronograms with only two tips. If numeric, it is used
as indices to subset chronograms object.}
}
\description{
subset chronograms for plotting densitree plots and phylo_all plots
}
