% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_node_ages.R
\name{plot_node_ages}
\alias{plot_node_ages}
\title{Add Node Age/Calibration Data to a Plotted Phylogeny}
\usage{
plot_node_ages(
  phy,
  cex_tips = 1,
  calibration_summary,
  color,
  pch = 20,
  cex_pch = 1
)
}
\arguments{
\item{phy}{A phylo object with branch length proportional to time}

\item{cex_tips}{A numeric value indicating **c**haracter **ex**pansion (i.e.,
size scaling factor) of `phy` tip labels. Default to 1.}

\item{calibration_summary}{An output of [datelife:::summary.matchedCalibrations()]}

\item{color}{A named vector of colors. Names must correspond to study references.
If vector is not named, colors will be recycled.}

\item{cex_pch}{A numeric value indicating **c**haracter **ex**pansion (i.e.,
size scaling factor) of node age point symbols defined by `pch`. Default to 1.}
}
\description{
Add Node Age/Calibration Data to a Plotted Phylogeny
}
