% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_phylo_all}
\alias{plot_phylo_all}
\title{Plot all Chronograms with Study Titles and Geochronological Axis}
\usage{
plot_phylo_all(
  chronograms,
  include = TRUE,
  individually = TRUE,
  write = "no",
  folder_name = "phylo_all",
  file_name = "chronogram",
  plot_type = "phyloch",
  max_depth,
  cex_tiplabels = graphics::par("cex"),
  cex_axislabel = graphics::par("cex"),
  cex_axis = graphics::par("cex"),
  cex_title = graphics::par("cex"),
  mai1,
  mai2,
  mai3,
  mai4,
  omi1,
  omi2,
  omi3,
  omi4,
  ...
)
}
\arguments{
\item{chronograms}{A list of chronograms as `multiPhylo` or as a plain `list` object.}

\item{include}{Boolean or numeric vector. Default to TRUE, keep all chronograms
in chronograms. If FALSE, exclude chronograms with only two tips. If numeric, it is used
as indices to subset chronograms object.}

\item{individually}{Boolean indicating if chronograms should be plotted one by one
or appended to the same file.}

\item{write}{A character vector of length 1 indicating the file extension to
write the plots to. Options are "pdf" or "png". Anything else will not write a file.}

\item{folder_name}{A character string indicating the name of the folder to write
the chronogram plot files to. Default to "phylo_all". Only relevant if `write = "png"` or `write = "pdf"`.}

\item{file_name}{A character string indicating a file name to write the chronogram
plots to. Relevant if `write = "png"` or `write = "pdf"`.}

\item{plot_type}{A character vector of length one indicating the type of
plot to generate. Options are:
\describe{
    \item{"ape"}{It uses the functions [ape::plot.phylo()] and [ape::axisPhylo()].}
    \item{"phyloch"}{It uses the functions [ape::plot.phylo()] and [phyloch::axisGeo()].}
    \item{"strap"}{It uses the function [strap::geoscalePhylo()] from the package [strap].}
    \item{"phytools"}{Not implemented yet. It will use functions from the package [phytools]}
    }}

\item{max_depth}{A numeric vector of length 1, indicating the upper limit of
the time scale on the x axis to be used on all plots. If none is provided, it is estimated
by getting the time depth of the oldest chronogram, adding 5 and rounding to the
closest 10. See details for more.}

\item{cex_tiplabels}{A numeric value indicating **c**haracter **ex**pansion (i.e.,
size scaling factor) of tip labels in `chronogram`. Default to value given by `graphics::par("cex")`.}

\item{cex_axislabel}{A numeric indicating character expansion for the time axis label. Default to value given by `graphics::par("cex")`.}

\item{cex_axis}{A numeric indicating character expansion for the axis. Default to value given by `graphics::par("cex")`.}

\item{cex_title}{A numeric indicating character expansion for the title. Default to value given by `graphics::par("cex")`.}

\item{mai1}{A numeric value indicating internal plot margin sizes
in inches. `mai4` is particularly important as it indicates the space needed for
plotting whole tip labels on the right margin of the plot.}

\item{mai2}{A numeric value indicating internal plot margin sizes
in inches. `mai4` is particularly important as it indicates the space needed for
plotting whole tip labels on the right margin of the plot.}

\item{mai3}{A numeric value indicating internal plot margin sizes
in inches. `mai4` is particularly important as it indicates the space needed for
plotting whole tip labels on the right margin of the plot.}

\item{mai4}{A numeric value indicating internal plot margin sizes
in inches. `mai4` is particularly important as it indicates the space needed for
plotting whole tip labels on the right margin of the plot.}

\item{omi1}{A numeric value indicating outter plot margin sizes in inches.}

\item{omi2}{A numeric value indicating outter plot margin sizes in inches.}

\item{omi3}{A numeric value indicating outter plot margin sizes in inches.}

\item{omi4}{A numeric value indicating outter plot margin sizes in inches.}

\item{...}{
  Arguments passed on to \code{\link[ape:plot.phylo]{ape::plot.phylo}}
  \describe{
    \item{\code{x}}{an object of class \code{"phylo"} or of class
    \code{"multiPhylo"}.}
    \item{\code{type}}{a character string specifying the type of phylogeny to be
    drawn; it must be one of "phylogram" (the default), "cladogram",
    "fan", "unrooted", "radial" or any unambiguous abbreviation of
    these.}
    \item{\code{use.edge.length}}{a logical indicating whether to use the edge
    lengths of the phylogeny to draw the branches (the default) or not
    (if \code{FALSE}). This option has no effect if the object of class
    \code{"phylo"} has no `edge.length' element.}
    \item{\code{node.pos}}{a numeric taking the value 1 or 2 which specifies the
    vertical position of the nodes with respect to their descendants. If
    \code{NULL} (the default), then the value is determined in relation
    to `type' and `use.edge.length' (see details).}
    \item{\code{show.tip.label}}{a logical indicating whether to show the tip
    labels on the phylogeny (defaults to \code{TRUE}, i.e. the labels
    are shown).}
    \item{\code{show.node.label}}{a logical indicating whether to show the node
    labels on the phylogeny (defaults to \code{FALSE}, i.e. the labels
    are not shown).}
    \item{\code{edge.color}}{a vector of mode character giving the colours used
    to draw the branches of the plotted phylogeny. These are taken to be
    in the same order than the component \code{edge} of \code{phy}. If
    fewer colours are given than the length of \code{edge}, then the
    colours are recycled.}
    \item{\code{edge.width}}{a numeric vector giving the width of the branches of
    the plotted phylogeny. These are taken to be in the same order than
    the component \code{edge} of \code{phy}. If fewer widths are given
    than the length of \code{edge}, then these are recycled.}
    \item{\code{edge.lty}}{same as the previous argument but for line types;
    1: plain, 2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash.}
    \item{\code{node.color}}{a vector of mode character giving the colours used
    to draw the perpendicular lines associated with each node of the
    plotted phylogeny. These are taken to be
    in the same order than the component \code{node} of \code{phy}. If
    fewer colours are given than the length of \code{node}, then the
    colours are recycled.}
    \item{\code{node.width}}{as the previous argument, but for line widths.}
    \item{\code{node.lty}}{as the previous argument, but for line types;
    1: plain, 2: dashed, 3: dotted, 4: dotdash, 5: longdash, 6: twodash.}
    \item{\code{font}}{an integer specifying the type of font for the labels: 1
    (plain text), 2 (bold), 3 (italic, the default), or 4 (bold
    italic).}
    \item{\code{cex}}{a numeric value giving the factor scaling of the tip and
    node labels (Character EXpansion). The default is to take the
    current value from the graphical parameters.}
    \item{\code{adj}}{a numeric specifying the justification of the text strings
    of the labels: 0 (left-justification), 0.5 (centering), or 1
    (right-justification). This option has no effect if \code{type =
      "unrooted"}. If \code{NULL} (the default) the value is set with
    respect of \code{direction} (see details).}
    \item{\code{srt}}{a numeric giving how much the labels are rotated in degrees
    (negative values are allowed resulting in clock-like rotation); the
    value has an effect respectively to the value of
    \code{direction} (see Examples). This option has no effect if
    \code{type = "unrooted"}.}
    \item{\code{no.margin}}{a logical. If \code{TRUE}, the margins are set to
    zero and the plot uses all the space of the device (note that this
    was the behaviour of \code{plot.phylo} up to version 0.2-1 of `ape'
    with no way to modify it by the user, at least easily).}
    \item{\code{root.edge}}{a logical indicating whether to draw the root edge
    (defaults to FALSE); this has no effect if `use.edge.length = FALSE'
    or if `type = "unrooted"'.}
    \item{\code{label.offset}}{a numeric giving the space between the nodes and
    the tips of the phylogeny and their corresponding labels. This
    option has no effect if \code{type = "unrooted"}.}
    \item{\code{underscore}}{a logical specifying whether the underscores in tip
    labels should be written as spaces (the default) or left as are (if
    \code{TRUE}).}
    \item{\code{x.lim}}{a numeric vector of length one or two giving the limit(s)
    of the x-axis. If \code{NULL}, this is computed with respect to
    various parameters such as the string lengths of the labels and the
    branch lengths. If a single value is given, this is taken as the
    upper limit.}
    \item{\code{y.lim}}{same than above for the y-axis.}
    \item{\code{direction}}{a character string specifying the direction of the
    tree. Four values are possible: "rightwards" (the default),
    "leftwards", "upwards", and "downwards".}
    \item{\code{lab4ut}}{(= labels for unrooted trees) a character string
    specifying the display of tip labels for unrooted trees (can be
    abbreviated): either \code{"horizontal"} where all labels are
    horizontal (the default if \code{type = "u"}), or \code{"axial"}
    where the labels are displayed in the axis of the corresponding
    terminal branches. This option has an effect if \code{type = "u"},
    \code{"f"}, or \code{"r"}.}
    \item{\code{tip.color}}{the colours used for the tip labels, eventually
    recycled (see examples).}
    \item{\code{plot}}{a logical controlling whether to draw the tree. If
    \code{FALSE}, the graphical device is set as if the tree was
    plotted, and the coordinates are saved as well.}
    \item{\code{rotate.tree}}{for "fan", "unrooted", or "radial" trees: the
    rotation of the whole tree in degrees (negative values are
    accepted).}
    \item{\code{open.angle}}{if \code{type = "f"} or \code{"r"}, the angle in
    degrees left blank. Use a non-zero value if you want to call
    \code{\link[ape]{axisPhylo}} after the tree is plotted.}
    \item{\code{node.depth}}{an integer value (1 or 2) used if branch lengths are
    not used to plot the tree; 1: the node depths are proportional to
    the number of tips descending from each node (the default and was the
    only possibility previously), 2: they are evenly spaced.}
    \item{\code{align.tip.label}}{a logical value or an integer. If \code{TRUE},
    the tips are aligned and dotted lines are drawn between the tips of
    the tree and the labels. If an integer, the tips are aligned and
    this gives the type of the lines (\code{lty}).}
  }}
}
\description{
Plot all Chronograms with Study Titles and Geochronological Axis
}
\details{
Currently, max_depth is obtained by default with `round(max(sapply(chronograms, function(x) max(ape::branching.times(x)))) + 5, digits = -1)`.
}
